Rename auth.rs to authentication.rs
