+++ crates/module-system/sov-modules-api/src/runtime/capabilities/authentication.rs
